changed the behavior of the front thread. Now the function process won't be killed after 6 seconds from the last request, it will only be killed if the monitor hasn't asked for its front in 6 seconds.

+++ computer/front_thread.py
@@ -73,7 +73,8 @@ class ProcessTimer(threading.Timer):
 # it kills the FunctionProcess, so it returns an End from the process queue, sets the front finished to True,
 # and removes the front from the live processes list.
 def process_timer(tuned_process: FunctionProcess, finished: Event, shared_info: SharedInfo):
-    shared_info.remove_alive(tuned_process.pid)
+    if shared_info.remove_alive(tuned_process.pid):
+        return
     finished.set()
     tuned_process.sabotage()
     tuned_process.kill()
